Fixed yet another issue with message relaying.
Ok, this one should finish it off. There was an issue where it tried to
read the element that was numbered as the last, but, being zero based,
the last is one shorter than what it was checking.

That should be fixed.

Oh, and I changed a few other things around. If I get a chance, I might
change them back. But iunno.