Fixed issue with colours in console
If a coloured message was sent through IRC, the colour would stay in the
terminal. Fixed by resetting the colour of the terminal to the
(sorta-)standardised white on black (sorry people that don't use that
kinda coloured terminal. Feel free to grab the source, modify it
according to the colours at
http://msdn.microsoft.com/en-us/library/system.consolecolor.aspx,
followed by a recompile from source).